checkpoint: renames "Config/utilities/DynamicConfig.ts" to "Config/Dynamic.ts"
