+++ src/features/TextToImage/Config/exports.ts
@@ -1,3 +1,3 @@
-export * as Config_Dynamic from './utilities/DynamicConfig';
+export * as Config_Dynamic from './Dynamic';
 export * as Config_Static from './utilities/StaticConfig';
 export * from './empty';
